Update URL paths to point to master

+++ cake-autorate_setup.sh
@@ -12,11 +12,13 @@ SRC_DIR="https://github.com/lynxthecat/cake-autorate/archive/refs/heads/"
 DOC_URL="https://github.com/lynxthecat/CAKE-autorate#installation-on-openwrt"
 BRANCH="testing"
 
-# Retrieve required packages
-printf "Running opkg update to update package lists:\n"
-opkg update
-printf "Installing bash, iputils-ping and fping packages:\n"
-opkg install bash iputils-ping fping
+# Retrieve required packages if not present
+if [ $(opkg list-installed | grep -E '^(bash|iputils-ping|fping) ' | wc -l) -ne 3 ]; then
+	printf "Running opkg update to update package lists:\n"
+	opkg update
+	printf "Installing bash, iputils-ping and fping packages:\n"
+	opkg install bash iputils-ping fping
+fi
 
 # Set up CAKE-autorate files
 # cd to the /root directory
@@ -63,9 +65,10 @@ mv "${tmp}/cake-autorate" /etc/init.d/
 chmod +x /etc/init.d/cake-autorate
 
 # Tell how to handle the config file - use old, or edit the new one
-printf '%s\n' "$editmsg"
+# shellcheck disable=SC2059
+printf "${editmsg}\n"
 
-printf '\n%s\n\n' "$(grep cake_autorate_version /root/cake-autorate/cake-autorate_config.primary.sh) successfully installed, but not yet running"
+printf '\n%s\n\n' "$(grep ^cake_autorate_version= /root/cake-autorate/cake-autorate_defaults.sh | cut -d= -f2 | cut -d'"' -f2) successfully installed, but not yet running"
 printf '%s\n' "Start the software manually with:"
 printf '%s\n' "   cd /root/cake-autorate; ./cake-autorate.sh"
 printf '%s\n' "Run as a service with:"
